Move User type definition to standalone module
