+++ web/src/types.ts
@@ -1,0 +1,7 @@
+export interface User{
+    id: string,
+    orcid: string;
+    name: string;
+    is_admin: boolean;
+    email?: string;
+}
